requirements.txt
h5py==2.6.0
Keras==2.0.4
matplotlib==2.0.0
numpy==1.12.1
Pillow==4.0.0
scikit-image==0.12.3
scipy==0.19.0
tensorflow-gpu==1.5.0
